route_optimizer/requirements-ai.txt
# Requirements with LLM/AI support
# Choose ONE of the following LLM providers:

# Option 1: OpenAI (Recommended - easiest to set up)
openai>=1.0.0

# Option 2: Anthropic Claude (Alternative)
# anthropic>=0.18.0

# Option 3: Local LLM with Ollama (Free, runs locally)
# No package needed - just install Ollama from https://ollama.ai

# Core requirements (required)
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
requests>=2.31.0


